+++ src/functions/matrix/hilbertMatrix.js
@@ -3,17 +3,16 @@ const LimeFunction = require('../../structs/function');
 
 /* ------------------------ division ------------------------ */
 
-// One matrix function class
-class LimeFunctionOneMatrix extends LimeFunction {
+// Hilbert matrix function class
+class LimeFunctionHilbertMatrix extends LimeFunction {
   // Constructor
   constructor(lime, mode) {
     // Super from function class
-    super(lime, { name: 'oneMatrix', mode });
+    super(lime, { name: 'hilbertMatrix', mode });
 
     // Operations
     this.operations.r = [
       'er(arg{int[1]})',
-      'er(arg{int[2]})',
     ];
 
     // Algorithms
@@ -22,18 +21,7 @@ class LimeFunctionOneMatrix extends LimeFunction {
       for (let j = 0; j < n; j++) {
         places.push([]);
         for (let i = 0; i < n; i++) {
-          places[j].push(this.lime.direct([1]));
-        }
-      }
-      step.rus(this.lime.build('matrix')(places));
-    });
-
-    this.algorithms.set('r(arg{int[2]})', (step) => {
-      const [m, n] = step.right.value; const places = [];
-      for (let j = 0; j < m; j++) {
-        places.push([]);
-        for (let i = 0; i < n; i++) {
-          places[j].push(this.lime.direct([1]));
+          places[j].push(this.lime.direct([1, '/', (i + j + 1)]));
         }
       }
       step.rus(this.lime.build('matrix')(places));
@@ -44,4 +32,4 @@ class LimeFunctionOneMatrix extends LimeFunction {
 /* ------------------------ division ------------------------ */
 
 // Export
-module.exports = LimeFunctionOneMatrix;
+module.exports = LimeFunctionHilbertMatrix;